Work around CACHE_LIFETIME_SAVED bug in smarty 3
In smarty 2 with CACHE_LIFETIME_SAVED you used to be able to set the
cache_lifetime just before calling fetch, and it would use that cache
lifetime, regardless of whether you had done stuff like is_cached
beforehand.
Now in smarty 3, whenever you do anything with a template, not just
fetch (so including isCached), it creates a Template class and stores it
internally, and reaccesses it later.  Good intentions, but the problem
is that it stores cache_lifetime on a per-template basis only once when
first creating the class, not when pulling up the stored class.  So in
other words, cache_lifetime has to be set before doing _anything_ with
the class at all, even checking if it isCached, in order for it to take
effect, which is a change in behavior from smarty 2.