Add clean url for history
Ideally the filename would be the the next parameter in the rest url.
The problem is that if the filename ends in php, the rest url ends up
being interpreted as an actual php file by the web server and the
request fails before the rewrite occurs.  For example:

/gitphp/projects/project.git/history/util.php

This tries to execute util.php in that nonexistent folder instead of
redirecting to the history of the file util.php.  If I can come up with
a better way to rewrite this I can revisit this problem in the future.