Bring back tag badges - fixes #6
This copy pasted code from the head list won’t really work for tags. Heads are easy in that they always point to a commit. Tags are more complicated in that they’re objects themselves, which then point to commits or even other tags. So you can’t just say the hash of the ref is the commit, you have to read the tag to dereference into the commit.